backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
mysqlclient  # ou psycopg2 se usar PostgreSQL
python-dotenv

